Added tests and minor documentation/fixes
